feat: 자동차 경주 진행
자동차 경주 기능 추가
- "실행 결과" 출력 후 경기 내용 출력
   - 각 자동차마다 0 ~ 9 사이의 무작위 값을 1회 구한 후 그 값이 4 이상인 경우에만 해당 자동차 1회 전진
     - 모든 자동차에서 1회의 무작위 값을 구한 후 "{자동차 이름} : {실행 결과}" 형식으로 출력
       - {실행 결과}에는 해당 자동차의 총 전진 성공 횟수만큼 "-"를 출력
     - 입력 받은 시도 횟수만큼 반복

+++ src/main/java/racingcar/InputReader.java
@@ -21,11 +21,11 @@ public class InputReader {
         return carNames;
     }
 
-    public long readTryCount() {
+    public int readTryCount() {
         System.out.println("시도할 횟수는 몇 회인가요?");
         String input = Console.readLine();
         validator.validateTryCount(input);
-        long tryCount = Long.parseLong(input);
+        int tryCount = Integer.parseInt(input);
         Console.close();
         return tryCount;
     }
